Metodo LicitarPinturas

+++ src/BelasArtes/DataLayer/Licitacao.cs
@@ -2,10 +2,10 @@ namespace DataLayer;
 
 public class Licitacao {
         public int? Id { get; set; }
-        public int Valor { get; set; }
+        public float Valor { get; set; }
         public string? EmailLicitador { get; set; } 
 
-        public Licitacao(int? id, int valor, string? emailLicitador){
+        public Licitacao(int? id, float valor, string? emailLicitador){
                 Id = id;
                 Valor = valor;
                 EmailLicitador = emailLicitador;
